pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-hub-dialogues"
version = "0.1.0"
description = "Offline topic analysis pipeline for OpenWebUI dialogue exports"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "pyarrow>=16.0",
  "PyYAML>=6.0",
  "matplotlib>=3.8",
  "scikit-learn>=1.5",
  "umap-learn>=0.5.6",
  "hdbscan>=0.8.36",
  "bertopic>=0.16.0",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
]

[tool.setuptools.packages.find]
include = ["src", "src.*"]

